Update patch set 6
Patch Set 6:

> Patch Set 6:
> 
> (1 comment)

Clement, the CL you are referring to was done as a performance improvement. Using this CL could possibly prevent the error but that doesn't mean it is wrong. It has been in Linux for over a year without seeing this problem, and as Carlos explained, it is documented in the ISA.

To me it is more suspicious that the failing code is only in places where it is reading a value from sched without being under control of sched.lock. If another block of code acquires this lock, and does several things, isn't the expectation that another goroutine won't try to read any of sched until the first block of code has finished all it needs to do?

Patch-set: 6