Update patch set 6
Patch Set 6:

> To me it is more suspicious that the failing code is only in places where it is reading a value from sched without being under control of sched.lock. If another block of code acquires this lock, and does several things, isn't the expectation that another goroutine won't try to read any of sched until the first block of code has finished all it needs to do?

At the time that those values are loaded, all the other P's are stopped, and the goroutine doing the check was woken via notetsleep which does a load-acquire.  There shouldn't be any way that the goroutine can see an out of date value that changes when doing an atomic load.  If there is a way, then something is wrong with notewakeup and notetsleep.  Or so it seems to me.  And, logic aside, it does seem to work on every other system.

If the problem isn't in the processor instructions that we are generating, then it almost has to be in the AIX implementation of semasleep and semawakeup, which on AIX are implemented as sem_timedwait and sem_post, but I don't see how those could fail to synchronize memory.

Patch-set: 6