Update patch set 1
Patch Set 1:

> Patch Set 1: Code-Review+2
> 
> Is there an issue for this? Or did you just notice it?

I "noticed" it doing the crazy inlining experiments, where "noticed" means that one of the benchmarks would hang, and I ran a binary search inlining level to find out the trigger, then saw the list of newly inlined functions, picked the obvious one (mstart or mstart1, don't recall) and stared at it for a while.

So it was a really hard notice, waiting to bite someone if they ever write a tiny piece of code that depends on this or they start searching for better inlining heuristics.

To Brad's remarks, I can try to write a test for this, but it might be a pain because these are internal-use-only intrinsics (and it's not like we have a test for every first-this-then-that piece of the compiler, we just got this one wrong by mistake).

Patch-set: 1
Reviewer: Gerrit User 7061 <7061@62eb7196-b449-3ce5-99f1-c037f21e1705>