Update patch set 2
Patch Set 2:

I know this is WIP and not review, but this would change the layout of slices, right? Any use of reflect.StringHeader => reflect.SliceHeader would break?

I know that both are documented with "It cannot be used safely or portably and its representation may change in a later release", but this type of conversion is a pretty well understood way to reduce garbage: https://github.com/search?l=Go&q=reflect+stringheader&type=Code.

Patch-set: 2
CC: Gerrit User 26898 <26898@62eb7196-b449-3ce5-99f1-c037f21e1705>