cmd/compile: always mark atColumn1 results as statements
In 31618, we end up comparing the is-stmt-ness of positions
to repurpose real instructions as inline marks. If the is-stmt-ness
doesn't match, we end up not being able to remove the inline mark.

Always use statement-full positions to do the matching, so we
always find a match if there is one.

TODO: this may redirect an inline mark to an instruction which
is not a statement. We may want to upgrade that instruction
to be statement-bearing in that case.

Fixes #31618

DO NOT REVIEW
DO NOT SUBMIT
(Not sure if the pos change is the right way to go yet)

Change-Id: Idaf39bdb32fa45238d5cd52973cadf4504f947d5