Update patch set 2
Patch Set 2:

> I know this is WIP and not review, but this would change the layout of slices, right? Any use of reflect.StringHeader => reflect.SliceHeader would break?

No. The nice thing about the approach this CL takes is that the memory layout of slice headers is unchanged and old reflect/unsafe code that accesses them will continue to work. This CL only changes how slice headers are represented in SSA form (i.e. once they have been loaded into registers).

As you point out we could change the memory layout since it is not subject to any guarantees but there would need to be a very big upside since there would be a lot of breakage.

> Would this help with #28942?

It should make pointer-only reslicing (as is done in your first example in that issue) a bit cheaper since we no longer need to update the cap of the slice in the loop. This probably won't reduce the net length of the generated machine code since we need to do a bit more work in the branch that we take when the length drops to 0 but it does, importantly, free up a register in the loop (and maybe reduces the length of the critical path too).

Eventually it might make sense for the compiler to split loops that have bounds checks. For example, the compiler could turn your code from this:

        for len(p) > 4 {
                // zero bounds checks.
                _ = p[0]
                _ = p[1]
                _ = p[2]
                _ = p[3]

                p = p[4:] // reslicing is expensive.
        }

Into this:

        for len(p) > 5 {
                // zero bounds checks.
                _ = p[0]
                _ = p[1]
                _ = p[2]
                _ = p[3]

                p = p[4:] // reslicing is cheap.
        }

        if len(p) > 4 {
                // lots of bounds checks, but only executed once
                _ = p[0]
                _ = p[1]
                _ = p[2]
                _ = p[3]

                p = p[4:] // reslicing is expensive.
        }

There are also other interesting experiments we could do around how we represent slices in SSA. We could have a start and end pointer for example. That would mean we wouldn't have to update the length on every iteration of a reslicing loop.

Patch-set: 2
Reviewer: Gerrit User 11990 <11990@62eb7196-b449-3ce5-99f1-c037f21e1705>
Work-in-progress: false