cmd/go: error if mismatched major versions causes use of pseudo-version
See suggestion 1 by thepudds in #31543.

When the major version of a module and the module path do not line up,
go automatically falls back to using a pseudo-version for that module.
This adds a warning indicating the module has an incorrect module path.

In the future, we should also handle when the major version of
the latest tag does not match the module path (e.g. when running
`go get github.com/pierrec/lz4`). Right now, we silently choose the
latest "good" version (with matching major versions), which is
unexpected.

Fixes #31498.

Change-Id: I880cfaa3cf51160d9b8d1bcb9e966e06369b50b4