cmd/go/internal/modload: make 'list -u' consider current pseudoversion
As pointed out by thepudds in #30634, the 'list -u' documentation states that the current version should be considered for upgrade:
The -u flag adds information about available upgrades. When the latest version of a given module is newer than the current one, list -u sets the Module's Update field to information about the newer module.

In go 1.12.4 (and current tip), an older version will be suggested as upgrade to a newer pseudo version.

Note that the test script included points at my own github repository and needs to be replaced before this CL is accepted.

Updates: #30634

Change-Id: If2c8887198884b8e7ccb3a604908065aa1f1878a