cmd/dist, misc/wasm: add WASM multibyte argv test
Adds a cmd/dist test as a follow-up of CL 173877.
The test requires Node.js to run: Node.js is available on
the GOOS=js builders and it also requires os/exec to run.
Ironically though, os/exec is not available on GOOS=js so
it'll run cross-compiled on other OSes with Node.js and
os/exec both available.

Updates #31645.

Change-Id: I1a4909e5b9fc7c1b79a94359b047bd5515991abb