misc/ios: add clang wrapper for self-hosted iOS builds
Similar to the existing clangwrap.sh for building and testing Go
on iOS from a macOS host, this CL adds a clang wrapper suitable
for self-hosted iOS builds.

The iOS kernel refuses to run unsigned binaries, which include shell
scripts. This wrapper is therefore a Go program that runs clang
with the appropriate flags. The wrapper attempts to detect whether
clang produced a binary, in which case it is signed using the ldid
utility.

With this wrapper we avoid embedding knowledge of iOS code signing
into into cmd/dist and cmd/go.

Updates #31722

Change-Id: I5ce33bede89009b4ae6bb2ef8480d929cedc82f1