cmd/compile: special case escape analysis for fmt.Printf etc
Do not submit, still benchmarking to see if it actually matters.
Proof of concept for special case treatment.

For real-world applications it is probably useful to
propagate the information "... arg is passed to fmt.Printf"
as a function/method attribute so that logging libraries
will benefit from this as well.

Change-Id: Iadc1a003b8aa2e20c6645612cfc2c1e09a3d77a2