Update patch set 1
Patch Set 1:

Can you explain why CL 122695 doesn’t work for #29964? Which specific fact is missing (after CL is applied) to eliminate the bound check?

It looks to me that CL 122695 is a better,
more generic and more maintainable way to extend prove. It’s a pity it doesn’t work as well as this CL which is really specific to the case at hand.

Patch-set: 1
Reviewer: Gerrit User 5340 <5340@62eb7196-b449-3ce5-99f1-c037f21e1705>