cmd/link: look up symbol names in pieces
DO NOT SUBMIT
DO NOT REVIEW

Updates the object file reader to pick apart symbol names into a
series of components (e.g. go."", foobar) and the pass those
components to the symbol name table, so as to increase symbol name
sharing/commoning. This patch WIP, doesn't work in some cases.

Note: Cherry's most recent change 174657 is going to render this
mostly moot: putting the packagepath in at compile time (in
combination with my previous change to use read-only memory to back
single-component symname strings) is going to get most of the benefits
of this patch series without most of the drawbacks.

Change-Id: Id92d624e36878f787b3ebb622fa1ed128519e0de