Update patch set 6
Patch Set 6:

> Patch Set 6:
> 
> Should we pull this all the way out into getQuery in internal/modget/get.go? If we don't have a check in modload.Query, I think getQuery is the next best place. QueryPackage and QueryPattern are wrappers over Query, and they're called in different situations.

I think QueryPackage should be fine. It is only ever used when we're trying to resolve an import path, and at that point it's important that the module path actually match the import path in question.

(The calls to Query that I'm worried about breaking are exactly the ones that are *not* through QueryPackage or QueryPattern, like the ones in fixVersion, addUpdate, moduleInfo, and listModules.)


> Checking in getQuery would mean we'd only get an error for 'go get'; an explicitly written tag in go.mod would still be silently replaced.

I don't think getQuery is quite the right place for this either, although I am a bit concerned about what happens if we run `go get -u` on a module that was previously tagged `+incompatible`: we shouldn't update a working +incompatible version to a point at which a newly-added `go.mod` file conflicts with the old import path.

Patch-set: 6