Update patch set 1
Patch Set 1: Code-Review-1

R=go1.14

Thank you for your PR. It is now May so the tree is closed for performance changes. Please feel free to continue to work on this PR, but it will not be reviewed until the 1.14 cycle opens later in the year. See https://golang.org/s/release

As this PR claims to improve the performance of the container package, please do some benchmark results using go test -bench=. -count=50 for the before and after values and include the output of benchstat old new in your commit message to show that this PR improved the performance of the container package.

Patch-set: 1
Reviewer: Gerrit User 5150 <5150@62eb7196-b449-3ce5-99f1-c037f21e1705>
Label: Code-Review=-1