Update patch set 2
Patch Set 2: Code-Review+2

This is an internal package, it's imported only by two packages that are both +build windows-only. So this should be be a low risk change to make. Feel free to wait for owner of this package if needed.

Patch-set: 2
Reviewer: Gerrit User 6005 <6005@62eb7196-b449-3ce5-99f1-c037f21e1705>
Label: Code-Review=+2