Update patch set 4
Patch Set 4:

> Patch Set 4:
> 
> > Patch Set 4:
> > 
> > (1 comment)
> > 
> > Is this for 1.13?
> 
> Not sure. This is what I used to debug issue #31915. With this I was able to spot problem easily. So I thought I just mail it.

That makes me tend to favor it, and the only worry would be is the output clear enough that when one of us uses it a year from now, how quickly will we figure out what we're looking at? I have started to wonder whether we should be doing a sort of quality-enhancing pass on all the stats and debugging bits that we've added to SSA, in order to make them more accessible/useful.

Patch-set: 4