Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> > Patch Set 1: Code-Review-1
> > 
> > (1 comment)
> 
> My motivation for adding this comes from this https://go-review.googlesource.com/c/go/+/174777
> 
> As we endup change static entries to dynamic entries. So it’s worth to add a note.

I don't think this would be the right place, though. We already have a comment about this in the OMAPLIT case, perhaps you could expand that?

Ideally the caller of o.expr(n) should not rely on *anything* about the status of n.Left. There might not even be an n.Left. For example, the OANDAND and OOROR cases return an n which is just a local variable. n.Left is then nil.

Patch-set: 1
Reviewer: Gerrit User 5200 <5200@62eb7196-b449-3ce5-99f1-c037f21e1705>