runtime: make TestPhysicalMemoryUtilization more robust
This change makes a few changes to TestPhysicalMemoryUtilization.

Firstly, it documents the 10% threshold currently used by the test which
was originally chosen arbitrarily, but now actually reflects
retainExtraPercent.

Secondly, it changes the overuse is calculated by using the heap goal
(NextGC) instead of HeapAlloc, to reflect the new condition that is
checked when scavenging during heap growth.

Thirdly, it changes the success condition to be when overuse is <=10%
as opposed to <10%, since the heap-growth scavenging mechanism may bring
usage down to exactly 10%.

Thanks to the fix to #32012, this means we can also significantly reduce
the memory footprint of this test, since it doesn't need to appear
"large" compared to the arena size anymore.

Fixes #32010.

Change-Id: Ia38130481e292f581da7fa3289c98c99dc5394ed