runtime: increase physical memory utilization threshold in test
This change increases the physical memory utilization threshold in the
PhysicalMemoryUtilization test from 10% to 15%. The increase reflects
the fact that eager heap scavenging is no longer guaranteed since the
scavenging policy is to retain a 10% overhead, and it's not guaranteed
to complete its goal by the time the test actually checks for memory
utilization. So, we pick retainExtraPercent+5 (10 + 5) % as the new
threshold to allow for some overhead instead of adding a sleep to the
test.

Fixes #32010.

Change-Id: Ia38130481e292f581da7fa3289c98c99dc5394ed