Update patch set 3
Patch Set 3:

Patch set 3 addresses the fault in the previous logic which was causing the tests to fail.  The previous patch was applying facts across the signed/unsigned domains inappropriately, particularly in the case of Eq/Neq Ops, where all facts are learned in both signed and unsigned domains regardless of the value's type.

Now, facts learned in domains which are not appropriate for the pre-extension value's type will no longer be repeated on the pre-extension value. E.g. When uint8 is zero-extended to int for indexing, signed domain facts learned about the int cannot always be applied to the uint8, despite the fact that uint8 -> int is a value-preserving zeroExt. 

This gets particularly hairy with bounds checks (as in the example above) and with Eg/Neq Ops (their facts are always applied in both the signed and unsigned domains). Some of the cross-domain inferences were appropriate, but some were not. Selectively applying the cross-domain inferences was difficult to reason about so I removed them all. 

But, looking through the extensions in isNonNegative recovers all of the facts required to remove the same bounds checks as patch 2, AND 17 new ones.

Patch-set: 3
Reviewer: Gerrit User 30765 <30765@62eb7196-b449-3ce5-99f1-c037f21e1705>