net/http: the comment of Server.IdleTimeout doesn't describe its behavior correctly
the current documentation is slightly incorrect. The earlier comment was

> IdleTimeout is the maximum amount of time to wait for the
> next request when keep-alives are enabled. If IdleTimeout
> is zero, the value of ReadTimeout is used. If both are
> zero, there is no timeout.

This got changed to the current version in this CL - https://go-review.googlesource.com/46434 as a fix for #20383.

The correct sequence is this. At the end of the current request, it sets a deadline for IdleTimeout or ReadTimeout, whichever is non zero. At the beginning of the next request, it sets a deadline for ReadHeaderTimeout or ReadTimeout, whichever is non zero. The missing piece in the documentation is the latter.

Fixes #32053

Change-Id: I43dd0252d1bcee6c29a8529abd84c84a49b2fba9
GitHub-Last-Rev: aaf83156eae21af355f832eff16d92c5bdf88759
GitHub-Pull-Request: golang/go#32164