Update patch set 5
Patch Set 5: Run-TryBot+1 Code-Review+1

I went through all three CLs about implementing this optimization, and I'm fairly sure that this is the best of the three. It is readable, it is maintainable (as it builds up on the already-existing practice of improving the facts within ft.update), and it has good results in terms of elided checks. IMO we should go through with this and abandon the other two.

Personally I would +2 this now, but I'm not qualified to judge whether it can be merged now or should wait. Austin, Josh, please let me know your thoughts.

(Also: what is the policy about +2-ing CLs that must wait? Can that be done or is it better to +2 only when they can actually be merged?)

Patch-set: 5
Reviewer: Gerrit User 5340 <5340@62eb7196-b449-3ce5-99f1-c037f21e1705>
Label: Code-Review=+1
Label: Run-TryBot=+1